requirements.txt
openai==1.12.0
pylint==3.0.3
black==24.2.0
radon==6.0.1
python-dotenv==1.0.1